requirements.txt
fastapi==0.95.0
uvicorn==0.18.3
python-dotenv==0.21.0
google-generativeai==0.1.0
pandas

